Removed return type-casting of CLib and OS abstraction functions.
* Should be left up to the caller.
* In NULL OS removed ineffective "#if <type>" checks.

+++ osal/null_os/inc/os_abstraction.h
@@ -47,91 +47,91 @@ typedef pltfm_thread_return_t (*PLTFM_THREAD_START_ROUTINE)(pltfm_thread_arg_t);
  */
 
 /*
-INT32 pltfm_thread_create(OUT pltfm_thread_t*,
+int pltfm_thread_create(OUT pltfm_thread_t*,
                            IN pltfm_thread_attr_t*,
                            IN PLTFM_THREAD_START_ROUTINE,
                            IN pltfm_thread_arg_t);
 */
-#define pltfm_thread_create(t, t_at, sr, t_ar)  ((INT32)null_thread_create((pltfm_thread_t*)(t),\
-                                                                           (pltfm_thread_attr_t*)(t_at),\
-                                                                           (PLTFM_THREAD_START_ROUTINE)(sr),\
-                                                                           (pltfm_thread_arg_t)(t_ar)))
+#define pltfm_thread_create(t, t_at, sr, t_ar)  null_thread_create((pltfm_thread_t*)(t),\
+                                                                   (pltfm_thread_attr_t*)(t_at),\
+                                                                   (PLTFM_THREAD_START_ROUTINE)(sr),\
+                                                                   (pltfm_thread_arg_t)(t_ar))
 
 /*
-INT32 pltfm_thread_cancel(IN pltfm_thread_t);
+int pltfm_thread_cancel(IN pltfm_thread_t);
 */
-#define pltfm_thread_cancel(t)              ((INT32)null_thread_cancel((pltfm_thread_t)(t)))
+#define pltfm_thread_cancel(t)              null_thread_cancel((pltfm_thread_t)(t))
 
 /*
-INT32 pltfm_thread_join(IN pltfm_thread_t,
+int pltfm_thread_join(IN pltfm_thread_t,
                         OUT void **retval);
 */
-#define pltfm_thread_join(t, retval)        ((INT32)null_thread_join((pltfm_thread_t)(t), (retval)))
+#define pltfm_thread_join(t, retval)        null_thread_join((pltfm_thread_t)(t), (retval))
 
 /*
-INT32 pltfm_thread_attr_init(OUT pltfm_thread_attr_t*);
+int pltfm_thread_attr_init(OUT pltfm_thread_attr_t*);
 */
-#define pltfm_thread_attr_init(t_at)        ((INT32)null_thread_attr_init((pltfm_thread_attr_t)(t_at)))
+#define pltfm_thread_attr_init(t_at)        null_thread_attr_init((pltfm_thread_attr_t)(t_at))
 
 /*
-INT32 pltfm_thread_attr_destroy(IN pltfm_thread_attr_t*);
+int pltfm_thread_attr_destroy(IN pltfm_thread_attr_t*);
 */
-#define pltfm_thread_attr_destroy(t_at)     ((INT32)null_thread_attr_destroy((pltfm_thread_attr_t)(t_at)))
+#define pltfm_thread_attr_destroy(t_at)     null_thread_attr_destroy((pltfm_thread_attr_t)(t_at))
 
 /*
-INT32 pltfm_thread_yield(void);
+int pltfm_thread_yield(void);
 */
-#define pltfm_thread_yield()                ((INT32)null_thread_yield())
+#define pltfm_thread_yield()                null_thread_yield()
 
 
 /* 
  * Task/Thread Synchronization Primitives
  */
 /*
-INT32 pltfm_mutex_init(OUT pltfm_mutex_t* ,
+int pltfm_mutex_init(OUT pltfm_mutex_t* ,
                        IN pltfm_mutexattr_t* );
 */
-#define pltfm_mutex_init(m, m_at)           ((INT32)null_thread_mutex_init((pltfm_mutex_t*)(m),\
-                                                                           (pltfm_mutexattr_t*)(m_at)))
+#define pltfm_mutex_init(m, m_at)           null_thread_mutex_init((pltfm_mutex_t*)(m),\
+                                                                   (pltfm_mutexattr_t*)(m_at))
 
 /*
-INT32 pltfm_mutex_destroy(IN pltfm_mutex_t* );
+int pltfm_mutex_destroy(IN pltfm_mutex_t* );
 */
-#define pltfm_mutex_destroy(m)              ((INT32)null_thread_mutex_destroy((pltfm_mutex_t*)(m)))
+#define pltfm_mutex_destroy(m)              null_thread_mutex_destroy((pltfm_mutex_t*)(m))
 
 /*
-INT32 pltfm_mutex_lock(INOUT pltfm_mutex_t* );
+int pltfm_mutex_lock(INOUT pltfm_mutex_t* );
 */
-#define pltfm_mutex_lock(m)                 ((INT32)null_thread_mutex_lock((pltfm_mutex_t*)(m)))
+#define pltfm_mutex_lock(m)                 null_thread_mutex_lock((pltfm_mutex_t*)(m))
 
 /*
-INT32 pltfm_mutex_unlock(INOUT pltfm_mutex_t* );
+int pltfm_mutex_unlock(INOUT pltfm_mutex_t* );
 */
-#define pltfm_mutex_unlock(m)               ((INT32)null_thread_mutex_unlock((pltfm_mutex_t*)(m)))
+#define pltfm_mutex_unlock(m)               null_thread_mutex_unlock((pltfm_mutex_t*)(m))
 
 /*
-INT32 pltfm_cond_init(OUT pltfm_cond_t* ,
+int pltfm_cond_init(OUT pltfm_cond_t* ,
                       IN pltfm_condattr_t* );
 */
-#define pltfm_cond_init(c, c_at)            ((INT32)null_thread_cond_init((pltfm_cond_t*)(c),\
-                                                                          (pltfm_condattr_t*)(c_at)))
+#define pltfm_cond_init(c, c_at)            null_thread_cond_init((pltfm_cond_t*)(c),\
+                                                                  (pltfm_condattr_t*)(c_at))
 
 /*
-INT32 pltfm_cond_destroy(OUT pltfm_cond_t* );
+int pltfm_cond_destroy(OUT pltfm_cond_t* );
 */
-#define pltfm_cond_destroy(c)               ((INT32)null_thread_cond_destroy((pltfm_cond_t*)(c)))
+#define pltfm_cond_destroy(c)               null_thread_cond_destroy((pltfm_cond_t*)(c))
 
 /*
-INT32 pltfm_cond_wait(INOUT pltfm_cond_t* ,
+int pltfm_cond_wait(INOUT pltfm_cond_t* ,
                       INOUT pltfm_mutex_t* );
 */
-#define pltfm_cond_wait(c, c_m)             ((INT32)null_thread_cond_wait((pltfm_cond_t*)(c),\
-                                                                          (pltfm_mutex_t*)(c_m)))
+#define pltfm_cond_wait(c, c_m)             null_thread_cond_wait((pltfm_cond_t*)(c),\
+                                                                  (pltfm_mutex_t*)(c_m))
 
 /*
-INT32 pltfm_cond_signal(INOUT pltfm_cond_t* );
+int pltfm_cond_signal(INOUT pltfm_cond_t* );
 */
-#define pltfm_cond_signal(c)                ((INT32)null_thread_cond_signal((pltfm_cond_t*)(c)))
+#define pltfm_cond_signal(c)                null_thread_cond_signal((pltfm_cond_t*)(c))
 
 
 /*
